+++ src/PommaLabs.Thrower/Properties/AssemblyInfo.cs
@@ -23,7 +23,6 @@
 
 using System;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Security;