+++ src/PommaLabs.Thrower/Reflection/PortableTypeInfo.cs
@@ -37,7 +37,7 @@ namespace PommaLabs.Thrower.Reflection
     [SecuritySafeCritical]
     public static class PortableTypeInfo
     {
-#if !(PORTABLE || NETSTD11 || NETSTD12)
+#if !(NETSTD10 || NETSTD11)
         internal const BindingFlags PublicAndPrivateInstanceFlags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance;
         internal const BindingFlags PublicInstanceFlags = BindingFlags.Public | BindingFlags.Instance;
 #endif
@@ -59,7 +59,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static Assembly GetTypeAssembly(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).Assembly;
 #else
             return type.Assembly;
@@ -87,7 +87,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<Attribute> GetCustomAttributes(Type type, bool inherit)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).GetCustomAttributes().ToArray();
 #else
             return type.GetCustomAttributes(inherit).Cast<Attribute>().ToArray();
@@ -105,7 +105,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<Attribute> GetCustomAttributes(MemberInfo memberInfo, bool inherit)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return memberInfo.GetCustomAttributes(inherit).ToArray();
 #else
             return memberInfo.GetCustomAttributes(inherit).Cast<Attribute>().ToArray();
@@ -120,7 +120,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<ConstructorInfo> GetConstructors(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(type).DeclaredConstructors.ToArray();
 #else
             return type.GetConstructors(PublicAndPrivateInstanceFlags);
@@ -143,7 +143,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static Type GetBaseType(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).BaseType;
 #else
             return type.BaseType;
@@ -158,7 +158,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static Type GetGenericTypeDefinition(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(type).GetGenericTypeDefinition();
 #else
             return type.GetGenericTypeDefinition();
@@ -173,7 +173,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<Type> GetGenericTypeArguments(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(type).GenericTypeArguments;
 #else
             return type.GetGenericArguments();
@@ -188,7 +188,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<Type> GetInterfaces(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(type).ImplementedInterfaces.ToArray();
 #else
             return type.GetInterfaces();
@@ -203,7 +203,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static IList<PropertyInfo> GetPublicProperties(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             var properties = new List<PropertyInfo>();
             while (type != null)
             {
@@ -240,7 +240,7 @@ namespace PommaLabs.Thrower.Reflection
             Raise.ArgumentNullException.IfIsNull(instance, nameof(instance), "Instance cannot be null");
             Raise.ArgumentException.IfIsNullOrWhiteSpace(propertyName, nameof(propertyName), "Given property cannot be null, empty or blank");
 
-#if !(PORTABLE || NETSTD11 || NETSTD12)
+#if !(NETSTD10 || NETSTD11)
             return FastMember.ObjectAccessor.Create(instance)[propertyName];
 #else
             var propertyInfo = GetPublicProperties(instance.GetType()).Single(p => p.Name == propertyName);
@@ -272,7 +272,7 @@ namespace PommaLabs.Thrower.Reflection
             return propertyInfo.GetValue(instance, EmptyObjectArray);
         }
 
-#if !(PORTABLE || NETSTD11 || NETSTD12)
+#if !(NETSTD10 || NETSTD11)
 
         /// <summary>
         ///   Gets the value of given property on given instance.
@@ -303,7 +303,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsAbstract(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsAbstract;
 #else
             return type.IsAbstract;
@@ -330,7 +330,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsClass(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsClass;
 #else
             return type.IsClass;
@@ -365,7 +365,7 @@ namespace PommaLabs.Thrower.Reflection
                 return false;
             }
 
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(obj.GetType()).IsAssignableFrom(IntrospectionExtensions.GetTypeInfo(type));
 #else
             return obj.GetType().IsAssignableFrom(type);
@@ -382,7 +382,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsEnum(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsEnum;
 #else
             return type.IsEnum;
@@ -409,7 +409,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsGenericType(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsGenericType;
 #else
             return type.IsGenericType;
@@ -436,7 +436,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsGenericTypeDefinition(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsGenericTypeDefinition;
 #else
             return type.IsGenericTypeDefinition;
@@ -467,7 +467,7 @@ namespace PommaLabs.Thrower.Reflection
                 return false;
             }
 
-#if (PORTABLE || NETSTD11 || NETSTD12)
+#if (NETSTD10 || NETSTD11)
             return IntrospectionExtensions.GetTypeInfo(type).IsAssignableFrom(IntrospectionExtensions.GetTypeInfo(obj.GetType()));
 #else
             return type.IsInstanceOfType(obj);
@@ -484,7 +484,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsInterface(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsInterface;
 #else
             return type.IsInterface;
@@ -511,7 +511,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsPrimitive(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsPrimitive;
 #else
             return type.IsPrimitive;
@@ -538,7 +538,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsValueType(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return IntrospectionExtensions.GetTypeInfo(type).IsValueType;
 #else
             return type.IsValueType;
@@ -565,7 +565,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsPublic(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return type.GetTypeInfo().IsPublic;
 #else
             return type.IsPublic;
@@ -588,7 +588,7 @@ namespace PommaLabs.Thrower.Reflection
         [MethodImpl(Raise.MethodImplOptions)]
         public static bool IsNestedPublic(Type type)
         {
-#if (PORTABLE || NETSTD11 || NETSTD12 || NETSTD13)
+#if (NETSTD10 || NETSTD11 || NETSTD13)
             return type.GetTypeInfo().IsNestedPublic;
 #else
             return type.IsNestedPublic;